requirements.txt
flask>=3.0
flask-wtf>=1.2
flask-limiter>=4.0
openpyxl>=3.1
python-dotenv>=1.0
werkzeug>=3.0
gunicorn>=21.2
